setup socketio server,controller,service, edit chat models, comment some console log
