+++ src/controllers/chatController.js
@@ -1,0 +1,3 @@
+exports.joinRoom = async (roomId, userId) => {
+    return;
+};
