+++ src/models/ChatRoom.js
@@ -1,31 +1,35 @@
 module.exports = (sequelize, DataTypes) => {
-  const ChatRoom = sequelize.define(
-    "ChatRoom",
-    {
-      roomTitle: {
-        type: DataTypes.STRING,
-      },
-    },
-    { underscored: true }
-  );
+    const ChatRoom = sequelize.define(
+        "ChatRoom",
+        {
+            roomTitle: {
+                type: DataTypes.STRING,
+            },
+            type: {
+                type: DataTypes.ENUM("DIRECT MESSAGE", "CHAT ROOM"),
+                defaultValue: "DIRECT MESSAGE",
+            },
+        },
+        { underscored: true }
+    );
 
-  ChatRoom.associate = (models) => {
-    ChatRoom.hasMany(models.Message, {
-      foreignKey: {
-        name: "chatRoomId",
-        allowNull: false,
-      },
-      onDelete: "RESTRICT",
-    });
+    ChatRoom.associate = (models) => {
+        ChatRoom.hasMany(models.Message, {
+            foreignKey: {
+                name: "chatRoomId",
+                allowNull: false,
+            },
+            onDelete: "RESTRICT",
+        });
 
-    ChatRoom.belongsTo(models.User, {
-      foreignKey: {
-        name: "senderId",
-        allowNull: false,
-      },
-      onDelete: "RESTRICT",
-    });
-  };
+        ChatRoom.hasMany(models.ChatMember, {
+            foreignKey: {
+                name: "chatRoomId",
+                allowNull: false,
+            },
+            onDelete: "RESTRICT",
+        });
+    };
 
-  return ChatRoom;
+    return ChatRoom;
 };
